fix: 抽离检查函数、同步markdownlint规则
