+++ packages/extension/package.json
@@ -17,24 +17,24 @@
     "test": "vscode-test"
   },
   "dependencies": {
-    "cspell-lib": "^9.1.2",
-    "js-yaml": "^4.1.0",
-    "markdownlint": "^0.38.0",
-    "webview-bridge": "*",
-    "shared": "*"
+    "checkers": "*",
+    "js-yaml": "4.1.0",
+    "markdownlint": "0.38.0",
+    "shared": "*",
+    "webview-bridge": "*"
   },
   "devDependencies": {
-    "@types/js-yaml": "^4.0.9",
-    "@types/mocha": "^10.0.10",
-    "@types/node": "^24.0.1",
-    "@types/vscode": "^1.75.0",
-    "@typescript-eslint/eslint-plugin": "^8.31.1",
-    "@typescript-eslint/parser": "^8.31.1",
-    "@vscode/test-cli": "^0.0.10",
-    "@vscode/test-electron": "^2.5.2",
-    "esbuild": "^0.25.3",
-    "eslint": "^9.25.1",
-    "npm-run-all": "^4.1.5",
-    "typescript": "^5.8.3"
+    "@types/js-yaml": "4.0.9",
+    "@types/mocha": "10.0.10",
+    "@types/node": "24.0.1",
+    "@types/vscode": "1.75.0",
+    "@typescript-eslint/eslint-plugin": "8.31.1",
+    "@typescript-eslint/parser": "8.31.1",
+    "@vscode/test-cli": "0.0.10",
+    "@vscode/test-electron": "2.5.2",
+    "esbuild": "0.25.3",
+    "eslint": "9.25.1",
+    "npm-run-all": "4.1.5",
+    "typescript": "5.8.3"
   }
 }
