+++ packages/extension/src/config/markdownlint.ts
@@ -5,6 +5,7 @@ export const MD_DEFAULT_CONFIG: Configuration = {
   default: true,
   MD003: { style: 'atx' },
   MD029: { style: 'ordered' },
+  MD004: false,
   MD007: false,
   MD009: false,
   MD013: false,
@@ -19,6 +20,7 @@ export const MD_DEFAULT_CONFIG: Configuration = {
   MD043: false,
   MD044: false,
   MD045: false,
+  MD046: false,
   MD048: false,
   MD049: false,
   MD050: false,
@@ -28,59 +30,4 @@ export const MD_DEFAULT_CONFIG: Configuration = {
   MD055: false,
   MD056: false,
   MD057: false,
-};
-
-// markdown-lint 报错中文描述
-export const MD_DESC: Record<string, string> = {
-  MD001: `标题等级只能每次扩大一个。{expected_actual}`,
-  MD003: `标题格式要统一，请去除标题下多余的 === 或 ---。{expected_actual}`,
-  MD004: `无序列表的格式要一致，列表前缀符号请使用同一种符号。{expected_actual}`,
-  MD005: `相同级别列表项的缩进要一致，请检测每一项缩进是否对齐。{expected_actual}`,
-  MD007: `无序列表缩进有误。{expected_actual}`,
-  MD009: `行尾有多余的空格。{expected_actual}`,
-  MD010: `请使用空格代替制表符缩进`,
-  MD011: `链接写法有误，请检查 [] 和 () 是否写反`,
-  MD012: `文档中有多余的空行，请删除多余的空行。{expected_actual}`,
-  MD013: `行数超过最大限制，请考虑换行。{expected_actual}`,
-  MD014: `命令前使用美元符号，会导致不显示输出`,
-  MD018: `标题#号后面缺少空格，请添加一个空格`,
-  MD019: `标题#号后面存在多余空格，请只保留一个空格`,
-  MD020: `标题#号前后请保留一个空格`,
-  MD021: `标题#号前后请只保留一个空格`,
-  MD022: `标题后请换行。{expected_actual}`,
-  MD023: `标题前有多余的空格或制表符，请删除多余的空格或制表符`,
-  MD024: `多个标题内容相同`,
-  MD025: `文档存在多个一级标题，请只保留一个`,
-  MD026: `标题的末尾存在标点符号，请删除多余的标点符号`,
-  MD027: `引用符号后面存在多余空格，请只保留一个空格。`,
-  MD028: `引用内容之间存在空白行，请删除空白行`,
-  MD029: `有序列表项前的数字请按顺序递增。{expected_actual}`,
-  MD030: `列表前缀符号后面请保留一个空格。{expected_actual}`,
-  MD031: `代码块前后应该换行`,
-  MD032: `列表前后应该换行`,
-  MD033: `文档内容不允许使用 html`,
-  MD034: `请勿直接书写链接，请使用 markdown 语法 <xxx> 或 [xxx](xxx)`,
-  MD035: `分隔线符号不一致，请使用同一种分隔线符号。{expected_actual}`,
-  MD036: `请勿使用加粗来代替标题`,
-  MD037: `加粗符号和被加粗文字之间不能有空格，但文字之间可以有空格，请删除多余的空格`,
-  MD038: `行内代码块和被包裹代码之间不能有空格，但代码之间可以有空格，请删除多余的空格`,
-  MD039: `链接名称和包围它的中括号之间不能有空格，但链接名称之间可以有空格，请删除多余的空格`,
-  MD040: `代码块应该指定一种编程语言`,
-  MD041: `文档的第一个非空行应该是一级标题`,
-  MD042: `链接地址为空`,
-  MD043: `标题编写需要按照指定的结构`,
-  MD044: `专有名称应该有正确的大写字母`,
-  MD045: `图片应该有替代的 alt 文本`,
-  MD046: `代码块应该保持一致的缩进格式`,
-  MD047: `文档末尾需要一个换行结束`,
-  MD048: `代码块使用语法不一致，请使用同一种代码块符号`,
-  MD049: `强调使用语法不一致，请使用同一种强调符号`,
-  MD050: `加粗使用语法不一致，请使用同一种加粗符号`,
-  MD051: `链路分片必须有效`,
-  MD052: `参考链接和图片应该使用已定义的标签`,
-  MD053: `链接和图像引用应用已定义的标签`,
-  MD055: `表格语法有误，请检查是否缺少 | 符号`,
-  MD056: `表格每一行的列数应该相同，请检查是否某一行列数不对`,
-  MD058: `表格前后应该换行`,
-  MD059: `链接文本应该是描述性的，禁止使用click here、here、link、more`,
 };
