+++ packages/extension/src/core/check/check-md-in-toc.ts
@@ -1,33 +1,7 @@
 import * as vscode from 'vscode';
-import path from 'path';
-import fs from 'fs';
-import { getYamlAsync } from 'shared';
 
-import { TocItem } from '@/@types/toc';
 import { isConfigEnabled } from '@/utils/common';
-
-/**
- * 判断 markdown 是否加入 _toc.yaml
- * @param {TocItem} toc toc item
- * @param {string} dirPath 当前目录
- * @param {string} mdPath markdown 路径
- * @returns {boolean} 加入返回 true，未加入返回 false
- */
-function isMdInToc(toc: TocItem, dirPath: string, mdPath: string) {
-  if (toc && typeof toc.href === 'string' && path.join(dirPath, toc.href).replace(/\\/g, '/') === mdPath) {
-    return true;
-  }
-
-  if (Array.isArray(toc.sections)) {
-    for (const item of toc.sections) {
-      if (isMdInToc(item, dirPath, mdPath)) {
-        return true;
-      }
-    }
-  }
-
-  return false;
-}
+import { execMdInTocCheck } from 'checkers';
 
 /**
  * 检查 markdown 是否加入 _toc.yaml
@@ -38,21 +12,8 @@ export async function checkMdInToc(document: vscode.TextDocument) {
     return;
   }
 
-  const mdPath = document.uri.fsPath.replace(/\\/g, '/');
-  let dirArr = mdPath.split('/');
-
-  while (dirArr.length) {
-    const dirPath = dirArr.join('/');
-    const tocPath = path.join(dirPath, '_toc.yaml');
-    if (fs.existsSync(tocPath)) {
-      const yamlObj = await getYamlAsync<TocItem>(tocPath);
-      if (isMdInToc(yamlObj, dirPath, mdPath)) {
-        return;
-      }
-    }
-
-    dirArr.pop();
+  const [isInToc] = await execMdInTocCheck(document.uri.fsPath);
+  if (!isInToc) {
+    vscode.window.showInformationMessage(`友情提示：${document.uri.path.split('/').pop()} 未加入到任何_toc.yaml文件`);
   }
-
-  vscode.window.showInformationMessage(`友情提示：${document.uri.path.split('/').pop()} 未加入到任何_toc.yaml文件`);
 }
