+++ packages/markdown/package.json
@@ -4,27 +4,27 @@
   "license": "MIT",
   "main": "src/markdown.ts",
   "dependencies": {
-    "@mdit-vue/plugin-component": "^2.1.4",
-    "@mdit-vue/plugin-frontmatter": "^2.1.4",
-    "@mdit-vue/plugin-headers": "^2.1.4",
-    "@mdit-vue/plugin-sfc": "^2.1.4",
-    "@mdit-vue/plugin-title": "^2.1.4",
-    "@mdit-vue/plugin-toc": "^2.1.4",
-    "@mdit-vue/shared": "^2.1.4",
-    "@shikijs/transformers": "^3.7.0",
-    "@shikijs/types": "^3.7.0",
-    "@types/markdown-it": "^14.1.2",
-    "@types/markdown-it-attrs": "^4.1.3",
-    "@types/markdown-it-container": "^2.0.10",
-    "@types/markdown-it-emoji": "^3.0.1",
-    "markdown-it": "^14.1.0",
-    "markdown-it-anchor": "^9.2.0",
-    "markdown-it-async": "^2.2.0",
-    "markdown-it-attrs": "^4.3.1",
-    "markdown-it-container": "^4.0.0",
-    "markdown-it-emoji": "^3.0.0",
-    "markdown-it-mathjax3": "^4.3.2",
-    "nanoid": "^5.1.5",
-    "shiki": "^3.7.0"
+    "@mdit-vue/plugin-component": "2.1.4",
+    "@mdit-vue/plugin-frontmatter": "2.1.4",
+    "@mdit-vue/plugin-headers": "2.1.4",
+    "@mdit-vue/plugin-sfc": "2.1.4",
+    "@mdit-vue/plugin-title": "2.1.4",
+    "@mdit-vue/plugin-toc": "2.1.4",
+    "@mdit-vue/shared": "2.1.4",
+    "@shikijs/transformers": "3.7.0",
+    "@shikijs/types": "3.7.0",
+    "@types/markdown-it": "14.1.2",
+    "@types/markdown-it-attrs": "4.1.3",
+    "@types/markdown-it-container": "2.0.10",
+    "@types/markdown-it-emoji": "3.0.1",
+    "markdown-it": "14.1.0",
+    "markdown-it-anchor": "9.2.0",
+    "markdown-it-async": "2.2.0",
+    "markdown-it-attrs": "4.3.1",
+    "markdown-it-container": "4.0.0",
+    "markdown-it-emoji": "3.0.0",
+    "markdown-it-mathjax3": "4.3.2",
+    "nanoid": "5.1.5",
+    "shiki": "3.7.0"
   }
 }
