+++ packages/shared/package.json
@@ -12,6 +12,6 @@
     }
   },
   "dependencies": {
-    "js-yaml": "^4.1.0"
+    "js-yaml": "4.1.0"
   }
 }
