+++ packages/webview-bridge/package.json
@@ -12,7 +12,7 @@
     }
   },
   "dependencies": {
-    "js-yaml": "^4.1.0",
+    "js-yaml": "4.1.0",
     "markdown": "*",
     "shared": "*"
   }
